EDITH/requirements.txt
pyttsx3
SpeechRecognition
keyboard
google-generativeai
python-decouple
requests
wikipedia
pywhatkit
